Refactor: move all OpenCV code in same package
Mini-code generation and display + QR-code outline moved to internal/detect package,
along with all other OpenCV-related code.
